// src/types/ProductType.ts
export interface Product {
    id: number;
    nombreProducto: string;
    dosis: string;
    intervalo: string;
    cultivos: string;
    casa: string;
    beneficios: string;
    modoDeUso: string;
    equipoDeAplicacion: string;
    informacionTecnica: string;
    categoria: string;
    nombreImagen: string;
  }
  